+++ app/src/main/java/com/example/android/nadris/util/BindingAdaptors.kt
@@ -1,11 +1,8 @@
 package com.example.android.nadris.util
 
-import android.graphics.drawable.Drawable
 import android.view.View
-import android.widget.ImageView
 import android.widget.TextView
 import androidx.databinding.BindingAdapter
-import com.example.android.nadris.R
 
 @BindingAdapter("isVisible")
 fun View.isVisible(visible:Boolean){
@@ -16,7 +13,7 @@ fun View.isVisible(visible:Boolean){
 }
 
 @BindingAdapter("stringRecourse", "formatterInput")
-fun TextView.TextWithFormat(stringRecourse:Int, /*vararg*/ formatterInput:Int){
+fun TextView.TextWithFormat(stringRecourse:Int, /*vararg*/ formatterInput:Long){
     this.text = String.format(context.getString(stringRecourse),formatterInput)
 }
 //@BindingAdapter("imageUrl", "error")
